Feature/add grpc service

+++ api/src/main/kotlin/com/example/helloworld/repository/HelloWorldRepository.kt
@@ -3,4 +3,4 @@ package com.example.helloworld.repository
 import com.example.helloworld.entity.HelloWorld
 import org.springframework.data.repository.CrudRepository
 
-interface HelloWorldRepository : CrudRepository<HelloWorld, Int>
+interface HelloWorldRepository : CrudRepository<HelloWorld, Long>
